final tweaks to get feat-adjustments in so we can support feats on objects
git-svn-id: svn+ssh://50.28.74.132/home/luminari/svn/luminari/myproject/trunk@4734 dda7f814-ccd3-4591-a6d7-1bd23e5e64af

+++ spell_prep.c
@@ -1508,7 +1508,7 @@ void begin_preparing(struct char_data *ch, int class) {
  *   character's skills
  *   character feats   */
 int compute_spells_prep_time(struct char_data *ch, int class, int circle, int domain) {
-  int prep_time = 0;
+  float prep_time = 0.0;
   int bonus_time = 0;
   int stat_bonus = 0;
   int level_bonus = 0;
